pyproject.toml
[tool.poetry]
name = "langgraph-agents"
version = "0.1.0"
description = "LangGraph with multiple agent frameworks integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
langchain = "^0.3.19"
langgraph = "^0.2.74"
autogen = "^0.7.5"
openai = "^1.64.0"
crewai = "^0.102.0"
langchain-community = "^0.3.18"
langchain-openai = "^0.3.7"
pydantic-ai = "^0.0.28"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
